% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cdisc_pilot.R
\docType{data}
\name{adae}
\alias{adae}
\title{ADAE}
\format{
Data frame with 32,139 features and 34 fields
}
\source{
<https://www.cdisc.org/sdtmadam-pilot-project>, downloaded 2020-05-15
}
\usage{
adae
}
\description{
Adverse Events Analysis Dataset from the CDISC Pilot Study
}
\keyword{datasets}
