% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{addBlock}
\alias{addBlock}
\title{Add Block to Block Data Object}
\usage{
addBlock(bd, variable, stat, dropdown, tpnt, df)
}
\arguments{
\item{bd}{A block data object}

\item{variable}{The parameter or field the statistic is based on}

\item{stat}{The statistic to be calculated}

\item{dropdown}{A subgroup on which the statistic is calculated (usually an AVISIT)}

\item{tpnt}{A time point on which the calculation is filtered}

\item{df}{The dataset the parameter or field is from}
}
\value{
The \code{bd} block data object with additional block
}
\description{
Add Block to Block Data Object
}
\examples{

datalist <- list(ADSL = tidyCDISC::adsl, ADVS = tidyCDISC::advs, 
                 ADAE = tidyCDISC::adae, ADLBC = tidyCDISC::adlbc)
bd <- createBlockdata(datalist)

\dontrun {
  addBlock(bd)
  bd
}

addBlock(bd, "DIABP", "MEAN", "ALL", "ALL")
bd
}
\keyword{table_blocks}
