% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cdisc_pilot.R
\docType{data}
\name{adlbc}
\alias{adlbc}
\title{ADLBC}
\format{
Data frame with 32,740 features and 58 fields
}
\source{
<https://www.cdisc.org/sdtmadam-pilot-project>, downloaded 2020-05-15
}
\usage{
adlbc
}
\description{
Laboratory Results Chemistry Analysis Dataset from the CDISC Pilot Study
}
\keyword{datasets}
