% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cdisc_pilot.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL}
\format{
Data frame with 254 features and 51 fields
}
\source{
<https://www.cdisc.org/sdtmadam-pilot-project>, downloaded 2020-05-15
}
\usage{
adsl
}
\description{
Subject Level Analysis Dataset from the CDISC Pilot Study
}
\keyword{datasets}
