% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_anova.R
\name{app_anova}
\alias{app_anova}
\alias{app_anova.default}
\alias{app_anova.ADAE}
\alias{app_anova.BDS}
\alias{app_anova.OCCDS}
\alias{app_anova.custom}
\title{Generate ANOVA using table generator blocks}
\usage{
app_anova(column, week, group, data)

\method{app_anova}{default}(column, week, group, data)

\method{app_anova}{ADAE}(column, week, group = NULL, data)

\method{app_anova}{BDS}(column, week, group = NULL, data)

\method{app_anova}{OCCDS}(column, week = NULL, group = NULL, data)

\method{app_anova}{custom}(column, week = NULL, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform ANOVA on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the ANOVA}

\item{data}{the data to use}
}
\value{
an ANOVA table of grouped variables

an ANOVA table of grouped variables

an ANOVA table of grouped variables
}
\description{
Generate ANOVA using table generator blocks

if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by

if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any
}
\seealso{
Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}

Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}

Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}

Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}

Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}

Other tableGen Functions: 
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}
}
\concept{tableGen Functions}
