% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_boxplot.R
\name{app_boxplot}
\alias{app_boxplot}
\title{IDEA boxplot}
\usage{
app_boxplot(data, yvar, group, value = NULL, points = FALSE)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar}{Selected y-axis}

\item{group}{Selected x-axis}

\item{value}{If yvar is a PARAMCD then the user must select 
AVAL, CHG, or BASE to be plotted on the y-axis}

\item{points}{\code{logical} whether to add a jitter to the plot}
}
\description{
Create boxplot using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
and filter the data by week
}
\seealso{
Other popExp Functions: 
\code{\link{app_heatmap}()},
\code{\link{app_lineplot}()},
\code{\link{app_spaghettiplot}()}
}
\concept{popExp Functions}
