% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_heatCorr.R
\name{app_heatmap}
\alias{app_heatmap}
\title{IDEA heatmap plot of endpoint correlations}
\usage{
app_heatmap(
  data,
  yvar_x,
  yvar_y,
  time,
  value = "AVAL",
  cor_mthd = "pearson",
  show_sig = F,
  sig_level = 0.05
)
}
\description{
Create a line plot with a time variable as the x-axis
and using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
to calculate the means. Lines are plotted by patient
}
\seealso{
Other popExp Functions: 
\code{\link{app_boxplot}()},
\code{\link{app_lineplot}()},
\code{\link{app_spaghettiplot}()}
}
\concept{popExp Functions}
