% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_km.R
\name{app_km_curve}
\alias{app_km_curve}
\title{IDEA Kaplan-Meier Curve}
\usage{
app_km_curve(
  data,
  yvar,
  resp_var,
  cnsr_var,
  group = "NONE",
  points = TRUE,
  ci = FALSE
)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar}{Selected xy-axis}

\item{group}{namne of grouping variable (categorical or factor)}

\item{points}{logical, whether to plot + symbols when patients censored}

\item{ci}{logical, whether the curve(s) should be accompanied with a 95\% CI}
}
\description{
Create scatter plot where if the variables are numeric then they
are plotted, and if they are PARAMCD's then a week and value 
must be selected for plotting.
}
\seealso{
Other popExp functions: 
\code{\link{app_scatterplot}()}
}
\concept{popExp functions}
\keyword{popEx}
