% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_max_freq.R
\name{app_max_freq}
\alias{app_max_freq}
\alias{app_max_freq.default}
\alias{app_max_freq.BDS}
\alias{app_max_freq.custom}
\title{Generate frequency where each subject is only counted once for the maximum
VAR}
\usage{
app_max_freq(column, group, data, totals)

\method{app_max_freq}{default}(column, group = NULL, data, totals)

\method{app_max_freq}{BDS}(column, group = NULL, data, totals)

\method{app_max_freq}{custom}(column, group, data, totals)
}
\arguments{
\item{column}{the variable to perform  stats on, this also contains
the class of the column based on the data file the column came from}

\item{group}{the groups to compare for the ANOVA}

\item{data}{the data to use}

\item{totals}{the totals data frame that contains denominator N's use when
calculating column percentages}
}
\value{
a frequency table of grouped variables

frequency table of ADSL column
}
\description{
Generate frequency where each subject is only counted once for the maximum
VAR

if ADSL supplied look for the column to take frequency of
and look for a grouping variable to group_by
if data is grouped add total column to the grouped data
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}

Other tableGen Functionss: 
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()}

Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
\concept{tableGen Functionss}
