% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{app_mean.BDS}
\alias{app_mean.BDS}
\title{if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any}
\usage{
\method{app_mean}{BDS}(column, week, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform summary statistics on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the summary statistics}

\item{data}{the data to use}
}
\value{
an summary statistic table of grouped variables
}
\description{
if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()}
}
\concept{tableGen Functions}
