% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{app_mean}
\alias{app_mean}
\alias{app_mean.ADAE}
\alias{app_mean.BDS}
\title{Generate summary statistics 
using table generator blocks}
\usage{
app_mean(column, week, group, data)

\method{app_mean}{ADAE}(column, week, group = NULL, data)

\method{app_mean}{BDS}(column, week, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform summary statistics on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the summary statistics}

\item{data}{the data to use}
}
\value{
an summary statistic table of grouped variables

an summary statistic table of grouped variables

an summary statistic table of grouped variables
}
\description{
Generate summary statistics 
using table generator blocks

if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by

if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean, if any
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()}

Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()}

Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()}
}
\concept{tableGen Functions}
\keyword{tabGen}
