% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_boxplot.R
\name{boxPlot_ui}
\alias{boxPlot_ui}
\title{Box Plot UI}
\usage{
boxPlot_ui(id, label = "box")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\description{
This module contains the widgets needed to create
a box plot
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
<<<<<<< HEAD
\code{\link{km_srv}()},
\code{\link{km_ui}()},
=======
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
>>>>>>> master
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
