% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExp_fct_buildEvents.R
\name{build_events}
\alias{build_events}
\title{Build Events Data Frame}
\usage{
build_events(
  input_checkbox,
  input_apply_filter,
  my_loaded_adams,
  my_datafile,
  my_filtered_dat,
  my_usubjid
)
}
\arguments{
\item{input_checkbox}{A character vector of event names selected. Some
possible event names include "Milestones", "Adverse Events", "Con Meds",
"Labs", "Medical History", etc...}

\item{input_apply_filter}{Logical; If \code{TRUE} then toggle data frame used to build
events to "my_filtered_dat", which is pre-filtered data frame from IDEAFilter}

\item{my_loaded_adams}{A character vector of available dataframes in
my_datafile}

\item{my_datafile}{A list of data frames}

\item{my_filtered_dat}{The output from IDEAFilter, a pre-filtered dataset}

\item{my_usubjid}{Character patient number, in the standard USUBJID format}
}
\value{
Data frame standardized to include events from various OCCDs class
  files
}
\description{
Gather various shiny inputs from OCCDs datasets (if they exist) and return a
single standardized events dataset
}
\seealso{
Other indvExp Functions: 
\code{\link{fnIndvExplVisits}()},
\code{\link{mod_indvExpPatEvents_server}()},
\code{\link{mod_indvExpPatVisits_server}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_server}()},
\code{\link{mod_indvExp_ui}()},
\code{\link{my_gg_color_hue}()},
\code{\link{org_df_events}()}
}
\concept{indvExp Functions}
