% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{CapStr}
\alias{CapStr}
\title{Capitalize the first letter of a string}
\usage{
CapStr(y)
}
\arguments{
\item{y}{the strings to capitalize}
}
\description{
Capitalize the first letter of a string
}
