% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{CDISC_mean.BDS}
\alias{CDISC_mean.BDS}
\title{if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any}
\usage{
\method{CDISC_mean}{BDS}(column, week, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform summary statistics on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the summary statistics}

\item{data}{the data to use}
}
\value{
an summary statistic table of grouped variables
}
\description{
if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
