% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{CDISC_methods}
\alias{CDISC_methods}
\title{Find the proper function to apply to 
each statistical and column block pairing
and use the metadata associated with 
each column block for the function's arguments}
\usage{
CDISC_methods(agg, column, week, group, data)
}
\arguments{
\item{agg}{the statistic to apply given the block name}

\item{column}{the column to apply that statistic too,
and class of the column dictated by the data frame it came from}

\item{week}{the week if needed for calculation}

\item{group}{whether to perform a group_by and if so by which column}

\item{data}{the dataset to perform all functions on}
}
\value{
the table corresponding to the proper function
to perform given the supplied column.
This is used within a map to apply to all blocks
inside the table generator module.
}
\description{
Find the proper function to apply to 
each statistical and column block pairing
and use the metadata associated with 
each column block for the function's arguments
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
