% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{check_params}
\alias{check_params}
\title{A function that checks if certain parameters exist in any dataframe within a list of dataframes}
\usage{
check_params(datafile, param_vector)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}

\item{param_vector}{character vector of params to search the list of dataframes for}
}
\description{
A function that checks if certain parameters exist in any dataframe within a list of dataframes
}
