% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{chem}
\alias{chem}
\title{Blood Chemistry PARAMCDs used to build STAN Table 41}
\format{
An object of class \code{character} of length 19.
}
\usage{
chem
}
\description{
Blood Chemistry PARAMCDs used to build STAN Table 41
}
\keyword{datasets}
