% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{cleanADAE}
\alias{cleanADAE}
\title{Function to clean and combine ADAE dataset with ADSL}
\usage{
cleanADAE(datafile, ADSL)
}
\arguments{
\item{datafile}{list of ADaM-ish data frames}

\item{ADSL}{A data frame which contains the ADSL data}
}
\value{
A cleaned ADAE data frame
}
\description{
Function to clean and combine ADAE dataset with ADSL
}
\details{
Finds the columns in the ADAE data frame, if it exists, in common with ADSL (besides USUBJID) and removes them from the ADAE so that the ADSL columns are used instead, then joins on USUBJID and re-orders the column names to match ADAE.
}
\keyword{tabGen_repro}
