% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{col_for_list_expr}
\alias{col_for_list_expr}
\title{GT Column Names}
\usage{
col_for_list_expr(nm, x)
}
\arguments{
\item{nm}{name}

\item{x}{columns

create the labels for each column using the total function
so the columns are now NAME N= X}

\item{dat}{data}
}
\description{
GT Column Names
}
\keyword{tabGen_repro}
