% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{combineBDS}
\alias{combineBDS}
\title{Combine BDS Data Frames}
\usage{
combineBDS(datafile, ADSL)
}
\arguments{
\item{datafile}{list of ADaM-ish data frames}

\item{ADSL}{A data frame which contains the ADSL data}
}
\value{
A data frame containing the BDS data bound by rows.
}
\description{
A function to combine all BDS data frames into one large data set.
}
\keyword{tabGen_repro}
