% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{combineData}
\alias{combineData}
\title{Function to merge the list of user supplied data frames}
\usage{
combineData(filepaths)
}
\arguments{
\item{datalist}{list of CDISC dataframes}
}
\description{
Function to merge the list of user supplied data frames
}
