% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{compareTables}
\alias{compareTables}
\title{Function to compare the SAS table to the IDEA output table}
\usage{
compareTables(sas_table, tg_table)
}
\arguments{
\item{sas_table}{SAS output}

\item{tg_table}{IDEA output}
}
\description{
Function to compare the SAS table to the IDEA output table
}
