% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{create_gt_table}
\alias{create_gt_table}
\title{Create a `gt` table object}
\usage{
create_gt_table(
  data,
  input_table_title,
  input_table_footnote,
  col_names,
  col_total,
  subtitle
)
}
\arguments{
\item{data}{The data frame used to create the table}

\item{input_table_title}{The Shiny input with the table title}

\item{input_table_footnote}{The Shiny input with the table footnote(s)}

\item{col_names}{A vector of column names}

\item{col_total}{A vector of column totals}

\item{subtitle}{The table subtitle}
}
\description{
Create a `gt` table object from the data that will then be used to generate
output in specific formats
}
\keyword{tabGen_repro}
