% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{createBlockdata}
\alias{createBlockdata}
\title{Create Block Data Object}
\usage{
createBlockdata(datalist)
}
\arguments{
\item{datalist}{A list of ADaM-ish datasets used to generate the table}
}
\value{
A block data object
}
\description{
Create Block Data Object
}
\examples{

datalist <- list(ADSL = tidyCDISC::adsl, ADVS = tidyCDISC::advs, 
                 ADAE = tidyCDISC::adae, ADLBC = tidyCDISC::adlbc)
bd <- createBlockdata(datalist)
bd
}
