% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{custom_class}
\alias{custom_class}
\title{Create custom S3 classes for each column block
based on the dataframe they came from}
\usage{
custom_class(x, df)
}
\arguments{
\item{x}{the name of each block}

\item{df}{the name of the data file
which will determine the class}
}
\description{
Create custom S3 classes for each column block
based on the dataframe they came from
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
