% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{data_to_use_str}
\alias{data_to_use_str}
\title{Function to clean and combine ADAE dataset with ADSL}
\usage{
data_to_use_str(x)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}
}
\description{
Function to clean and combine ADAE dataset with ADSL
}
