% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{delim_expand_rows}
\alias{delim_expand_rows}
\title{Expands a collapsed column into multiple rows}
\usage{
delim_expand_rows(data, sep)
}
\arguments{
\item{data}{either a tg_table or sas_table dataframe}

\item{sep}{a delimiter to search for when expanding rows}
}
\description{
Function that searches for a particular delimiter in data columns that would
then separate values on the left of the right into their own row for
increased machine readability.
}
