% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{deliver_avals_by}
\alias{deliver_avals_by}
\title{Deliver 'avals_by' value to the parent frame/ env}
\usage{
deliver_avals_by(x)
}
\arguments{
\item{x}{a value to assign to `avals_by` in the parent env}
}
\description{
Deliver 'avals_by' value to the parent frame/ env
}
\keyword{internal}
