% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{dropArea}
\alias{dropArea}
\title{The entire drop area of the blocks,
both statistical and columns}
\usage{
dropArea(name, id, ulid, class, styles, col)
}
\arguments{
\item{name}{the name of the drop area}

\item{id}{the id of the drop area}

\item{ulid}{the id of the list item}

\item{class}{the class of the drop area 
so that blocks can be dropped here but not others}

\item{styles}{the aesthetics to apply}

\item{col}{the width of the drop area}
}
\description{
The entire drop area of the blocks,
both statistical and columns
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
