% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{error_handler}
\alias{error_handler}
\title{Error handler}
\usage{
error_handler(e)
}
\arguments{
\item{e}{error message}
}
\value{
Error message
}
\description{
Function to handle errors based on specific method used
}
