% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example.R
\docType{data}
\name{example_dat1}
\alias{example_dat1}
\title{Example Data Set 1}
\format{
A list with 3 elements:
\describe{
  \item{AE}{data frame, pre-processed AE dataset}
  \item{BDS}{data frame, pre-processed BDS dataset}
  \item{totals}{data frame, contains totals by grouping variable for pre-processed data}
}
}
\usage{
example_dat1
}
\description{
Pre-processed data for purposes of demonstrating \link[tidyCDISC]{app_methods}.
}
\keyword{datasets}
