% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example.R
\docType{data}
\name{example_dat2}
\alias{example_dat2}
\title{Example Data Set 2}
\format{
A list with 3 elements:
\describe{
  \item{TG_table}{data frame, pre-processed \code{gt} table object with basic column names}
  \item{col_names}{vector, the column names}
  \item{col_totals}{vector, totals corresponding to each column}
}
}
\usage{
example_dat2
}
\description{
Pre-processed data for the purposes of demonstrating \link[tidyCDISC]{col_for_list_expr}.
}
\keyword{datasets}
