% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{filterData}
\alias{filterData}
\title{Function to filter rows from combined data}
\usage{
filterData(datafile, input_filter_df, filter_code)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}
}
\description{
Function to filter rows from combined data
}
