% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExpPatVisits_fct_plot.R
\name{fnIndvExplVisits}
\alias{fnIndvExplVisits}
\title{Plot Individual Explorer Visits by Param}
\usage{
fnIndvExplVisits(
  watermark = FALSE,
  graph_output = "plotly",
  bds_data,
  usubjid,
  input_plot_hor,
  input_visit_var,
  input_plot_param,
  input_plot_adam,
  input_overlay_events,
  vline_dat,
  vv_dy_name
)
}
\arguments{
\item{watermark}{If \code{TRUE}, then include a watermark on the output plot}

\item{graph_output}{A character string specifying either "plotly" or
"ggplot2"}

\item{bds_data}{A character string containg the name of the BDS Dataset}

\item{usubjid}{A character string containing Patient number in the form of
USUBJID standards}

\item{input_plot_hor}{A character string containing variable name to have
horizontal line plotted}

\item{input_visit_var}{A character string containing the visit variable name}

\item{input_plot_param}{A character string containing the PARAMCD name /
Value from which to plot AVAL values}

\item{input_plot_adam}{A character string containing the ADaM dataset name}

\item{input_overlay_events}{A character vector containing the names of
patient events to plot}

\item{vline_dat}{The vline data frame that contains x-intercept values for the
corrresponding events selected to be overlain on the plot}

\item{vv_dy_name}{TA character vector containing the name of the visit
variable(s)}
}
\value{
If graph_output is \code{plotly}, then a plotly object, else if
  \code{ggplot} then a ggplot2 object
}
\description{
Gather various shiny inputs and return either a ggplot or plotly object. This
graph's purpose is to plot AVAL against a visit variable for any PARAMCD,
establishing one part of patient narrative found on the Individual Explorer
Tab. The graph allows many features to be toggled to enhance the data
presented in the plot, including plotting horizontal lines for screening &
baseline measurements (when available) and overlaying pertinent OCCDs events
(certain milestones, adverse events, and/or con meds) that have occurred
during the patients journey.
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{mod_indvExpPatEvents_server}()},
\code{\link{mod_indvExpPatVisits_server}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_server}()},
\code{\link{mod_indvExp_ui}()},
\code{\link{my_gg_color_hue}()}
}
\concept{indvExp Functions}
