% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComply_fct_helpers.R
\name{gather_reqs}
\alias{gather_reqs}
\title{Gather Violated Rules}
\usage{
gather_reqs(
  input,
  output,
  session,
  disp_type = c("error", "warn"),
  datalist = reactive(NULL),
  all_df_rules = list(error = c(""), warn = c("")),
  expl_rules = list(list(error = c(""), warn = c(""))),
  df_incl_rules = list(list(error = c(""), warn = c("")))
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{disp_type}{Check for rules that would result in an error (required
variables) or warning (recommended variables)}

\item{datalist}{A reactive list of data frames (from the data upload module)}

\item{all_df_rules}{A double-nested list: inner list is named list of
variables names that should result in \code{error} or \code{warn} if
variables do not exist or are missing for ANY data frame. Outter list is
unnamed.}

\item{expl_rules}{A double-nested list: outter list is a named list
dataframes. Inner list (for each data frame) contains named lists of
variables names that should result in \code{error} or \code{warn} if
variables do not exist or are missing for the specified parent data frame}

\item{df_incl_rules}{A double-nested list: outter list is a named list
variable names used to identify a particular class of data frame. For
example, the variable name PARAMCD would id BDS class data frames, and the
inner list would contain a named lists of variables names that should
result in \code{error} or \code{warn} if variables do not exist or are
missing for the implied parent data frame.}
}
\value{
A list of dataframes that are compliant with the rules in addition to
  a gt and data frame object which explain which files violate the rules (if
  any)
}
\description{
Gather applicable Rules into a usable output (gt & df objects) to be included
in the pop-up modal when an uploaded data frame violates said rule(s). This
function also validates the rules entered to verify they are in the expected
format.
}
\seealso{
Other dataComply Functions: 
\code{\link{gather_rules}()},
\code{\link{mod_dataComplyRules_server}()},
\code{\link{mod_dataComplyRules_ui}()},
\code{\link{mod_dataComply_server}()},
\code{\link{mod_dataComply_ui}()}
}
\concept{dataComply Functions}
