% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComplyRules_fct_helpers.R
\name{gather_rules}
\alias{gather_rules}
\title{Gather Rules For Help UI}
\usage{
gather_rules(
  input,
  output,
  session,
  all_df_rules = list(list(error = c(""), warn = c(""))),
  expl_rules = list(list(error = c(""), warn = c(""))),
  df_incl_rules = list(list(error = c(""), warn = c("")))
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{all_df_rules}{A double-nested list: inner list is named list of
variables names that should result in \code{error} or \code{warn} if
variables do not exist or are missing for ANY data frame. Outter list is
unnamed.}

\item{expl_rules}{A double-nested list: outter list is a named list
dataframes. Inner list (for each data frame) contains named lists of
variables names that should result in \code{error} or \code{warn} if
variables do not exist or are missing for the specified parent data frame}

\item{df_incl_rules}{A double-nested list: outter list is a named list
variable names used to identify a particular class of data frame. For
example, the variable name PARAMCD would id BDS class data frames, and the
inner list would contain a named lists of variables names that should
result in \code{error} or \code{warn} if variables do not exist or are
missing for the implied parent data frame.}
}
\value{
An shiny tagList
}
\description{
Gather Rules into a shiny tagList to be included in the UI for "help"
  purposes, conditionally containing different required and/or recommended rule
  sets (if they exist).
}
\seealso{
Other dataComply Functions: 
\code{\link{gather_reqs}()},
\code{\link{mod_dataComplyRules_server}()},
\code{\link{mod_dataComplyRules_ui}()},
\code{\link{mod_dataComply_server}()},
\code{\link{mod_dataComply_ui}()}
}
\concept{dataComply Functions}
