% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExp_utils_helpers.R
\docType{data}
\name{guide_ind_exp_visits_adlb_olay}
\alias{guide_ind_exp_visits_adlb_olay}
\title{Individual Explorer Cicerone Guide For Patient Visits tab when ADLB loaded &
'dy' visit var selected, plus overlain events selected}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 15.
}
\usage{
guide_ind_exp_visits_adlb_olay
}
\value{
a cicerone r6 object
}
\description{
This object is used within the table generator Individual Explorer to add
help text to the various fields using a help buttom
}
\keyword{datasets}
