% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_utils.R
\docType{data}
\name{guide_popex}
\alias{guide_popex}
\title{Population Explorer Cicerone Guide}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 13.
}
\usage{
guide_popex
}
\value{
a cicerone r6 object
}
\description{
This object is used within the Population Explorer to add
help text to the various fields upon click of a help buttom
}
\keyword{datasets}
