% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_utils.R
\docType{data}
\name{guide_popex_sel_adv}
\alias{guide_popex_sel_adv}
\title{Population Explorer Cicerone Guide with Advanced Filtering}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 15.
}
\usage{
guide_popex_sel_adv
}
\value{
a cicerone r6 object
}
\description{
This object is used within the Population Explorer to add help text to the
various fields upon click of a help buttom
}
\keyword{datasets}
