% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataUpload_fct_hasColumn.R
\name{hasColumn}
\alias{hasColumn}
\title{Check whether a column is found in a data set}
\usage{
hasColumn(columnName, data)
}
\arguments{
\item{columnName}{The column to look for.}

\item{data}{the data.frame to search.}
}
\value{
logical scalar. TRUE if the column is found. FALSE otherwise
}
\description{
Checks whether a specified column is found in a specified data set
}
\keyword{internal}
