% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{hema}
\alias{hema}
\title{Hematology PARAMCDs used to build STAN Table 41}
\format{
An object of class \code{character} of length 14.
}
\usage{
hema
}
\description{
Hematology PARAMCDs used to build STAN Table 41
}
\keyword{datasets}
