% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_anova.R
\name{IDEA_anova}
\alias{IDEA_anova}
\alias{IDEA_anova.default}
\alias{IDEA_anova.ADAE}
\alias{IDEA_anova.BDS}
\alias{IDEA_anova.OCCDS}
\alias{IDEA_anova.custom}
\title{Generate ANOVA using table generator blocks}
\usage{
IDEA_anova(column, week, group, data)

\method{IDEA_anova}{default}(column, week, group, data)

\method{IDEA_anova}{ADAE}(column, week, group = NULL, data)

\method{IDEA_anova}{BDS}(column, week, group = NULL, data)

\method{IDEA_anova}{OCCDS}(column, week = NULL, group = NULL, data)

\method{IDEA_anova}{custom}(column, week = NULL, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform ANOVA on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the ANOVA}

\item{data}{the data to use}
}
\value{
an ANOVA table of grouped variables

an ANOVA table of grouped variables

an ANOVA table of grouped variables
}
\description{
Generate ANOVA using table generator blocks

if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by

if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
