% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct.R
\name{IDEA_boxplot}
\alias{IDEA_boxplot}
\title{IDEA boxplot}
\usage{
IDEA_boxplot(data, yvar, group, value = NULL, points = FALSE)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar}{Selected y-axis}

\item{group}{Selected x-axis}

\item{value}{If yvar is a PARAMCD then the user must select 
AVAL, CHG, or BASE to be plotted on the y-axis}

\item{points}{\code{logical} whether to add a jitter to the plot}
}
\description{
Create boxplot using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
and filter the data by week
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
