% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_freq.R
\name{IDEA_freq}
\alias{IDEA_freq}
\alias{IDEA_freq.default}
\alias{IDEA_freq.ADSL}
\alias{IDEA_freq.BDS}
\alias{IDEA_freq.OCCDS}
\alias{IDEA_freq.custom}
\title{Generate frequency of categorical variables
using table generator blocks}
\usage{
IDEA_freq(column, week, group, data)

\method{IDEA_freq}{default}(column, week, group, data)

\method{IDEA_freq}{ADSL}(column, week, group = NULL, data)

\method{IDEA_freq}{BDS}(column, week, group = NULL, data)

\method{IDEA_freq}{OCCDS}(column, week = NULL, group, data)

\method{IDEA_freq}{custom}(column, week = NULL, group, data)
}
\arguments{
\item{column}{the variable to perform frequency stats on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the ANOVA}

\item{data}{the data to use}
}
\value{
a frequency table of grouped variables

frequency table of ADSL column
}
\description{
Generate frequency of categorical variables
using table generator blocks

if ADSL supplied look for the column to take frequency of
and look for a grouping variable to group_by
}
