% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_heatCorr.R
\name{IDEA_heatmap}
\alias{IDEA_heatmap}
\title{IDEA line plot}
\usage{
IDEA_heatmap(data, yvar_x, yvar_y, time, value = "AVAL", cor_mthd = "pearson")
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{time}{Selected x-axis constained to time dependent columns}

\item{value}{If yvar is a PARAMCD then the user must select 
AVAL, CHG, or BASE to be plotted on the y-axis}

\item{yvar}{Selected y-axis}
}
\description{
Create a line plot with a time variable as the x-axis
and using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
to calculate the means. Lines are plotted by patient
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{heatmap_srv}()},
\code{\link{heatmap_ui}()},
\code{\link{km_srv}()},
\code{\link{km_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
