% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_line.R
\name{IDEA_lineplot}
\alias{IDEA_lineplot}
\title{IDEA line plot}
\usage{
IDEA_lineplot(
  data,
  yvar,
  time,
  value = NULL,
  separate = "NONE",
  color = "NONE",
  err_bars = FALSE,
  label_points = FALSE,
  gtxt_x_pos = "middle",
  gtxt_y_pos = "top",
  add_vert,
  vert_x_int,
  add_hor,
  hor_y_int
)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar}{Selected y-axis}

\item{time}{Selected x-axis constained to time dependent columns}

\item{value}{If yvar is a PARAMCD then the user must select 
AVAL, CHG, or BASE to be plotted on the y-axis}
}
\description{
Create a line plot with a time variable as the x-axis
and using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
to calculate the means. Lines are plotted by patient
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{km_srv}()},
\code{\link{km_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
