% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_max_freq.R
\name{IDEA_max_freq}
\alias{IDEA_max_freq}
\alias{IDEA_max_freq.default}
\alias{IDEA_max_freq.BDS}
\alias{IDEA_max_freq.custom}
\title{Generate frequency where each subject is only counted once for the maximum
VAR}
\usage{
IDEA_max_freq(column, group, data)

\method{IDEA_max_freq}{default}(column, group = NULL, data)

\method{IDEA_max_freq}{BDS}(column, group = NULL, data)

\method{IDEA_max_freq}{custom}(column, group, data)
}
\arguments{
\item{column}{the variable to perform  stats on, this also contains
the class of the column based on the data file the column came from}

\item{group}{the groups to compare for the ANOVA}

\item{data}{the data to use}
}
\value{
a frequency table of grouped variables

frequency table of ADSL column
}
\description{
Generate frequency where each subject is only counted once for the maximum
VAR

if ADSL supplied look for the column to take frequency of
and look for a grouping variable to group_by
if data is grouped add total column to the grouped data
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functionss: 
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()}

Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}

Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
\concept{tableGen Functionss}
