% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{IDEA_mean.ADSL}
\alias{IDEA_mean.ADSL}
\title{if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by}
\usage{
\method{IDEA_mean}{ADSL}(column, week, group = NULL, data)
}
\value{
an summary statistic table of grouped variables
}
\description{
if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by
}
