% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{IDEA_mean.BDS}
\alias{IDEA_mean.BDS}
\title{if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any}
\usage{
\method{IDEA_mean}{BDS}(column, week, group = NULL, data)
}
\value{
an summary statistic table of grouped variables
}
\description{
if BDS filter by paramcd and week
We need to calculate the difference in N for this
and report missing values from the mean if any
}
