% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{IDEA_mean}
\alias{IDEA_mean}
\alias{IDEA_mean.default}
\alias{IDEA_mean.OCCDS}
\alias{IDEA_mean.custom}
\title{Generate summary statistics 
using table generator blocks}
\usage{
IDEA_mean(column, week, group, data)

\method{IDEA_mean}{default}(column, week, group, data)

\method{IDEA_mean}{OCCDS}(column, week = NULL, group, data)

\method{IDEA_mean}{custom}(column, week = NULL, group, data)
}
\arguments{
\item{column}{the variable to perform summary statistics on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the summary statistics}

\item{data}{the data to use}
}
\value{
an summary statistic table of grouped variables
}
\description{
Generate summary statistics 
using table generator blocks
}
