% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct.R
\name{IDEA_spaghettiplot}
\alias{IDEA_spaghettiplot}
\title{IDEA spaghetti plot}
\usage{
IDEA_spaghettiplot(data, yvar, time, value = NULL)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar}{Selected y-axis}

\item{time}{Selected x-axis constained to time dependent columns}

\item{value}{If yvar is a PARAMCD then the user must select 
AVAL, CHG, or BASE to be plotted on the y-axis}
}
\description{
Create a spaghetti plot with a time dependent variable as the x-axis
and using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value. 
Lines are plotted by patient
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
