% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExp_utils_helpers.R
\docType{data}
\name{ind_exp_events}
\alias{ind_exp_events}
\title{Individual Explorer Cicerone Guide For Analyzing Patient Event Data}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 13.
}
\usage{
ind_exp_events
}
\description{
This object is used within the table generator Individual Explorer to add
help text to the various fields using a help buttom
}
\keyword{datasets}
