% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_line.R
\name{linePlot_ui}
\alias{linePlot_ui}
\title{Spaghetti Plot UI}
\usage{
linePlot_ui(id, label = "line")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\description{
This module contains the widgets needed to create
a spaghetti plot
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
