% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{machine_readable}
\alias{machine_readable}
\title{Organize SAS or tg table into a machine readable format for comparing}
\usage{
machine_readable(data)
}
\arguments{
\item{data}{a sas or tg table}
}
\description{
Organize SAS or tg table into a machine readable format for comparing
}
