% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{make_machine_readable}
\alias{make_machine_readable}
\title{Organize SAS or tg table into a machine readable format for comparing}
\usage{
make_machine_readable(data, keep_orig_ids = FALSE)
}
\arguments{
\item{data}{a sas or tg table}
}
\description{
Organize SAS or tg table into a machine readable format for comparing
}
