% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_meanSummary.R
\name{mean_summary}
\alias{mean_summary}
\title{the \code{mean_summary} function is used by
both the basic summary statistics block
and the change from baseline block}
\usage{
mean_summary(.data, to_count)
}
\arguments{
\item{.data}{is the data to pass to the summary statics}

\item{to_count}{is either AVAL or CHG}
}
\description{
the \code{mean_summary} function is used by
both the basic summary statistics block
and the change from baseline block
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
