% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{mk_rep_seq_id}
\alias{mk_rep_seq_id}
\title{Make repeated sequence id}
\usage{
mk_rep_seq_id(x)
}
\arguments{
\item{x}{a numeric vector of repeated sequence}
}
\description{
Make an "ID vector" that identifies groups of repeated sequences in another vector.
}
\examples{
mk_rep_seq_id(rep(1:5,2))

}
