% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComply.R
\name{mod_dataComply_server}
\alias{mod_dataComply_server}
\title{dataComply Server Function}
\usage{
mod_dataComply_server(
  input,
  output,
  session,
  datalist = reactive(NULL),
  dismissErrBttn = T
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{datalist}{A reactive list of data frames (from the upload module)}

\item{dismissErrBttn}{If \code{TRUE} (the default) then the 'dismiss' button
  will appear on the error modal. If \code{FALSE}, the user will not be able
  to escape the modal. Instead they will have to reload the app.

  DO NOT REMOVE.}
}
\value{
A list of dataframes that are compliant with the rules
}
\description{
A module that will interface with a list of data frames and either (I)
display an error if needed variables don't exist and stop them from
proceeding or (II) warn the user if if some columns are missing that are
vital for the app to make sense, but they can continue if they wish.
}
\concept{dataComply Functions}
