% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComply.R
\name{mod_dataComply_ui}
\alias{mod_dataComply_ui}
\title{dataComply UI Function}
\usage{
mod_dataComply_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module Not in use
}
