% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComplyRules.R
\name{mod_dataComplyRules_server}
\alias{mod_dataComplyRules_server}
\title{dataComplyRules Server Function}
\usage{
mod_dataComplyRules_server(
  input,
  output,
  session,
  modal_title = "ADaM-ish Upload Rules"
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{modal_title}{A text string to be displayed as the modal's title}
}
\description{
Situates output from gather_rules function in
"mod_dataComplyRules_fct_helpers.R" to a modal condiational on clicking
"help".
}
\seealso{
Other dataComply Functions: 
\code{\link{gather_reqs}()},
\code{\link{gather_rules}()},
\code{\link{mod_dataComplyRules_ui}()},
\code{\link{mod_dataComply_server}()},
\code{\link{mod_dataComply_ui}()}
}
\concept{dataComply Functions}
