% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComplyRules.R
\name{mod_dataComplyRules_ui}
\alias{mod_dataComplyRules_ui}
\title{dataComplyRules UI Function}
\usage{
mod_dataComplyRules_ui(id, showRules = T)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{showRules}{If TRUE, the UI will contain the returned shiny tagList will
contain a help button and respective modal when clicked. If FALSE, the
dataComply module will not include a help button to display rules.}
}
\value{
An shiny tagList
}
\description{
A shiny Module.
}
\seealso{
Other dataComply Functions: 
\code{\link{gather_reqs}()},
\code{\link{gather_rules}()},
\code{\link{mod_dataComplyRules_server}()},
\code{\link{mod_dataComply_server}()},
\code{\link{mod_dataComply_ui}()}
}
\concept{dataComply Functions}
