% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataUpload.R
\name{mod_dataUpload_server}
\alias{mod_dataUpload_server}
\title{dataUpload Server Function stores 
the uploaded data as a list and 
is exported to be used in other modules}
\usage{
mod_dataUpload_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\value{
a list of dataframes 
\code{dd$dataframe} 
to be used in other modules
}
\description{
dataUpload Server Function stores 
the uploaded data as a list and 
is exported to be used in other modules
}
