% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExp_srv.R
\name{mod_indvExp_server}
\alias{mod_indvExp_server}
\title{indvExp Server Function}
\usage{
mod_indvExp_server(input, output, session, datafile)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{datafile}{A list of dataframes}
}
\value{
character vector of loaded adams and a filtered dataframe to populate mod_indvExpPat module
}
\description{
Prepare Individual Explorer Tab with server side code which sets up some
initial properties on the tab. Specifically, the user will be able to
pre-filter the loaded data in order to select a subject using IDEAFilter.
That filtered data will be sent to additional modules for ingestion.
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_ui}()},
\code{\link{my_gg_color_hue}()}
}
\concept{indvExp Functions}
