% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExp_ui.R
\name{mod_indvExp_ui}
\alias{mod_indvExp_ui}
\title{indvExp UI Function}
\usage{
mod_indvExp_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that compiles the user interface for the Individual Explorer
Tab. In summary, the UI's main components include an IDEAFilter widget for
prefiltering, a brief demographic table, and another collection of tabset
panels that divide out the Patient Events (leveraging a timevis object that
plots dates on a timeline) from patient events by Parameter graphic (plotly).
Note: well panels encompass any user input fields
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{fnIndvExplVisits}()},
\code{\link{mod_indvExpPatEvents_server}()},
\code{\link{mod_indvExpPatVisits_server}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_server}()},
\code{\link{my_gg_color_hue}()},
\code{\link{org_df_events}()}
}
\concept{indvExp Functions}
