% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExpPat.R
\name{mod_indvExpPat_server}
\alias{mod_indvExpPat_server}
\title{indvExpPat Server Function}
\usage{
mod_indvExpPat_server(
  input,
  output,
  session,
  datafile,
  loaded_adams,
  filtered_dat
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{datafile}{A list of dataframes}

\item{loaded_adams}{a character vector of loaded adam datasets}

\item{filtered_dat}{a filtered dataframe containing USUBJID}
}
\value{
character string containing a USUBJID
}
\description{
Prepare Individual Explorer Tab with content post selection of a UBSUBJID. Specifically,
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{mod_indvExp_server}()},
\code{\link{mod_indvExp_ui}()},
\code{\link{my_gg_color_hue}()}
}
\concept{indvExp Functions}
