% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indvExpPatVisits.R
\name{mod_indvExpPatVisits_server}
\alias{mod_indvExpPatVisits_server}
\title{indvExpPatVisits Server Function}
\usage{
mod_indvExpPatVisits_server(
  input,
  output,
  session,
  datafile,
  loaded_adams,
  usubjid,
  filtered_dat
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{datafile}{A list of dataframes}

\item{loaded_adams}{A character vector of loaded adam datasets}

\item{usubjid}{A Character string containing a USUBJID}

\item{filtered_dat}{A IDEAFilter output data frame containing USUBJID}
}
\description{
Prepare Individual Explorer Tab's Visits subtab with content. Specifically,
creating some headers, build events dataset, Output a DT & plotly Object.
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{fnIndvExplVisits}()},
\code{\link{mod_indvExpPatEvents_server}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_server}()},
\code{\link{mod_indvExp_ui}()},
\code{\link{my_gg_color_hue}()}
}
\concept{indvExp Functions}
