% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp.R
\name{mod_popExp_server}
\alias{mod_popExp_server}
\title{popExp Server Function}
\usage{
mod_popExp_server(input, output, session, datafile)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{datafile}{A list of dataframes}
}
\description{
Prepare Individual Explorer Tab with some of the basics
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
<<<<<<< HEAD
\code{\link{km_srv}()},
\code{\link{km_ui}()},
=======
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
>>>>>>> master
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
