% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_ui.R
\name{mod_popExp_ui}
\alias{mod_popExp_ui}
\title{popExp UI Function}
\usage{
mod_popExp_ui(id, label = "Population Explorer")
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{label}{Name of module}
}
\description{
UI for population explorer: the filtering widget
as well as radio buttons for different plots. These radio buttons
are used to toggle between the child modules of each plot type, and 
a conditional panel of widgets based on the plot type
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
