% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen.R
\name{mod_tableGen_server}
\alias{mod_tableGen_server}
\title{tableGen Server Function}
\usage{
mod_tableGen_server(input, output, session, datafile = reactive(NULL))
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{datafile}{all uploaded datafiles from the dataImport module}
}
\value{
the grouped blocks for columns
to be used in the tableGen UI
}
\description{
tableGen Server Function
}
