% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_ui.R
\name{mod_tableGen_ui}
\alias{mod_tableGen_ui}
\title{tableGen UI Function}
\usage{
mod_tableGen_ui(id)
}
\arguments{
\item{id}{internal parameters for {shiny}.}
}
\description{
drag and drop table generator module
}
\details{
The UI is comprised of a drop zone
of both statistical and column blocks
as well as a means to filter and group the data.
The output is a gt table
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
