% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{my_gg_color_hue}
\alias{my_gg_color_hue}
\title{My GG Color Hue}
\usage{
my_gg_color_hue(n)
}
\arguments{
\item{n}{An output reactive dataframe from IDEAFilter}
}
\value{
character vector of hex colors of length n
}
\description{
Grab specific colors from the ggplot2 default color palette of length n
}
\seealso{
Other indvExp Functions: 
\code{\link{build_events}()},
\code{\link{fnIndvExplVisits}()},
\code{\link{mod_indvExpPatEvents_server}()},
\code{\link{mod_indvExpPatVisits_server}()},
\code{\link{mod_indvExpPat_server}()},
\code{\link{mod_indvExp_server}()},
\code{\link{mod_indvExp_ui}()}
}
\concept{indvExp Functions}
