% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{numeric_stan_table}
\alias{numeric_stan_table}
\title{Function to clean and combine ADAE dataset with ADSL}
\usage{
numeric_stan_table(input_recipe)
}
\arguments{
\item{input_recipe}{The shiny input that keeps track of the recipe selected}
}
\description{
Function to clean and combine ADAE dataset with ADSL
}
