% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{prep_bds}
\alias{prep_bds}
\title{Function to bind data rows from the list of user supplied data frames}
\usage{
prep_bds(datafile, ADSL)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}

\item{ADSL}{A dataframe which contains the ADSL data}
}
\description{
Function to bind data rows from the list of user supplied data frames
}
\keyword{tabGen_repro}
