% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{prep_tg_table}
\alias{prep_tg_table}
\title{Organize Table Generator table into a format for comparing}
\usage{
prep_tg_table(
  data,
  machine_readable = TRUE,
  keep_orig_ids = FALSE,
  rm_desc_col = FALSE,
  generic_colnames = "NONE"
)
}
\arguments{
\item{data}{the tg_table dataframe, output from tidyCDISC}

\item{machine_readable}{a logical; should the table be prepared for optimal
machine readability; that is, should cells with multiple values be pivoted
to new rows}

\item{keep_orig_ids}{a logical; if machine_readability is desired, do you
want to keep old id's that keep track of the values original position in
the table}

\item{rm_desc_col}{a logical; should the description column be removed from
the table (since they never match between base and compare)}

\item{generic_colnames}{a logical; Should the column names of the table
generator output be generalized to be col0 - colX col99, where x is the
number of groups in the 'group by' variable and col99 is the total column}
}
\description{
Organize Table Generator table into a format for comparing
}
\keyword{tabGen_compare}
