% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{pretty_blocks}
\alias{pretty_blocks}
\title{Table Generator Pretty Block lookup table}
\format{
A data farm with 8 rows and 2 variables:
\describe{
  \item{Pattern}{standard statistical name}
  \item{Replacement}{"pretty" name for table output}
}
}
\usage{
pretty_blocks
}
\description{
A dataset used within the table generator module
to add pretty names for each stat block when displayed in the table
}
\keyword{tabGen_repro}
