% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{recipe_column}
\alias{recipe_column}
\title{function to change the selected input of input$COLUMN based on the recipe block}
\usage{
recipe_column(column)
}
\arguments{
\item{column}{the selected column to group by}
}
\description{
function to change the selected input of input$COLUMN based on the recipe block
}
