% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct.R
\name{refact}
\alias{refact}
\title{Refactor common variables by their "N" counterparts}
\usage{
refact(data, varc, varn)
}
\arguments{
\item{data}{A data frames}

\item{varc}{A character vector of variables names with character values}

\item{varn}{A character vector of variables names with numeric values}
}
\description{
Neat trick to first set the selected choice to "0", then to character(0)
}
