% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{removeBlock}
\alias{removeBlock}
\title{Remove Block(s) from Block Data Object}
\usage{
removeBlock(bd, x)
}
\arguments{
\item{bd}{A block data object}

\item{x}{vector specifying elements to remove from block data object}
}
\value{
The \code{bd} block data object with additional block
}
\description{
Remove Block(s) from Block Data Object
}
