% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{revert_temp_colnames}
\alias{revert_temp_colnames}
\title{Revert temporary column names back to the original; use with temp_col_rename}
\usage{
revert_temp_colnames(dat, orig_grp_names)
}
\arguments{
\item{dat}{a data frame}

\item{orig_grp_names}{a character vector of original col names}
}
\description{
Revert temporary column names back to the original; use with temp_col_rename
}
