% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{rowArea}
\alias{rowArea}
\title{The entire row area of accordion blocks
generated within the table generator server}
\usage{
rowArea(bins, col)
}
\arguments{
\item{bins}{is the number of accordions to create}

\item{col}{is the column with to apply}
}
\description{
The entire row area of accordion blocks
generated within the table generator server
}
