% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{rowBlock}
\alias{rowBlock}
\title{Create a single draggable block for each column}
\usage{
rowBlock(name, classname)
}
\arguments{
\item{name}{the column name}

\item{classname}{the dataset of the column}
}
\description{
Create a single draggable block for each column
}
