% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{rowPallete}
\alias{rowPallete}
\title{Create an accordion dropdown for each file}
\usage{
rowPallete(data)
}
\arguments{
\item{data}{the file uploaded}
}
\description{
Create an accordion dropdown for each file
}
