% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\value{
No return value, called to run the application.
}
\description{
Run the Shiny Application
}
