% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_scatter.R
\name{scatterPlot_srv}
\alias{scatterPlot_srv}
\title{Scatter Plot Server Function}
\usage{
scatterPlot_srv(input, output, session, data)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{data}{The combined dataframe from population explorer}
}
\value{
ggplot object
}
\description{
Using the widgets from the scatter plot UI
create a ggplot object which is returned to the 
parent Population Explorer module
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{km_srv}()},
\code{\link{km_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
