% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_scatter.R
\name{scatterPlot_ui}
\alias{scatterPlot_ui}
\title{Scatter Plot UI}
\usage{
scatterPlot_ui(id, label = "scatter")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\description{
This module contains the widgets needed to create
a scatter plot
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{scatterPlot_srv}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
