% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{setGroup}
\alias{setGroup}
\title{Set the title for the table object}
\usage{
setGroup(bd, group_by)
}
\arguments{
\item{bd}{A block data object}

\item{group_by}{A field to group the table by}
}
\value{
The \code{bd} block data object with updated group by field
}
\description{
Set the title for the table object
}
\keyword{table_blocks}
