% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{setTitle}
\alias{setTitle}
\title{Set the title for the table object}
\usage{
setTitle(bd, title)
}
\arguments{
\item{bd}{A block data object}

\item{title}{The title for the table}
}
\value{
The \code{bd} block data object with supplied title
}
\description{
Set the title for the table object
}
\keyword{table_blocks}
