% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adam.R
\name{show_adamish_alert}
\alias{show_adamish_alert}
\title{Show ADaM Compatibility Alert}
\usage{
show_adamish_alert(validation, session)
}
\arguments{
\item{validation}{A list returned from `validate_adamish()`}

\item{session}{The Shiny session (typically `session` from server)}
}
\description{
Display a shinyalert-style popup based on the results of `validate_adamish()`
}
