% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_spaghetti.R
\name{spaghettiPlot_ui}
\alias{spaghettiPlot_ui}
\title{Spaghetti Plot UI}
\usage{
spaghettiPlot_ui(id, label = "spaghetti")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\description{
This module contains the widgets needed to create
a spaghetti plot
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()}
}
\concept{popExp Functions}
