% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{std_err}
\alias{std_err}
\title{Standard Error Calculation}
\usage{
std_err(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{logical, should NA's be removed? Defaults to FALSE}
}
\value{
numeric, representing the standard error
}
\description{
Calculates the square root of variance divided by n
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_max_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_nested_freq}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
