% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{std_footnote}
\alias{std_footnote}
\title{Create Standard Footnotes for TG Table}
\usage{
std_footnote(data, source)
}
\arguments{
\item{data}{The `gt` table object to append the footnote}

\item{source}{The source of the data in the table}
}
\value{
a `gt` object
}
\description{
Creates a footnote with a source on the left and date run on the right.
}
\keyword{tabGen_repro}
