% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adam.R
\name{suggest_adam_column_mapping}
\alias{suggest_adam_column_mapping}
\title{Suggest Column Mappings for ADaM}
\usage{
suggest_adam_column_mapping(
  df,
  validation,
  string_distance = FALSE,
  session = NULL,
  input_ids = NULL
)
}
\arguments{
\item{df}{A data.frame to evaluate}

\item{validation}{A list returned from `validate_adamish()`}

\item{string_distance}{Logical. Use string distance matching to provide suggestions, requires `stringdist` package}

\item{session}{A Shiny session object (optional).}

\item{input_ids}{A named list of input IDs corresponding to ADaM variable names (optional)}
}
\value{
A named list with best-guess mappings for missing variables
}
\description{
Based on validation results, attempt to suggest which columns to rename to standard ADaM variable names.
}
