% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adam.R
\name{suggest_adam_domain}
\alias{suggest_adam_domain}
\title{Suggest ADaM Domain}
\usage{
suggest_adam_domain(df)
}
\arguments{
\item{df}{A data.frame to evaluate.}
}
\value{
A character string fo the most likely domain name (e.g., "ADLB"), or NULL.
}
\description{
Heuristically suggests the most likely ADaM domain for a dataset
based on the presence of required variables.
}
\keyword{internal}
