% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{table_blocks}
\alias{table_blocks}
\title{import R6}
\description{
import R6

import R6
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datalist}}{A list of ADaM-ish datasets used to generate the table}

\item{\code{all_rows}}{A list of parameters/fields from the datalist with descriptions}

\item{\code{blocks}}{A data frame containing the block data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-table_blocks-new}{\code{table_blocks$new()}}
\item \href{#method-table_blocks-print}{\code{table_blocks$print()}}
\item \href{#method-table_blocks-add_block}{\code{table_blocks$add_block()}}
\item \href{#method-table_blocks-remove_block}{\code{table_blocks$remove_block()}}
\item \href{#method-table_blocks-clone}{\code{table_blocks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_blocks-new"></a>}}
\if{latex}{\out{\hypertarget{method-table_blocks-new}{}}}
\subsection{Method \code{new()}}{
Create a new block data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_blocks$new(datalist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datalist}}{A list of ADaM-ish datasets used to generate the table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_blocks-print"></a>}}
\if{latex}{\out{\hypertarget{method-table_blocks-print}{}}}
\subsection{Method \code{print()}}{
Print the data frame containing the blocks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_blocks$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_blocks-add_block"></a>}}
\if{latex}{\out{\hypertarget{method-table_blocks-add_block}{}}}
\subsection{Method \code{add_block()}}{
Add block to the block data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_blocks$add_block(variable, stat, dropdown, tpnt, df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{The parameter or field the statistic is based on}

\item{\code{stat}}{The statistic to be calculated}

\item{\code{dropdown}}{A subgroup on which the statistic is calculated (usually an AVISIT)}

\item{\code{tpnt}}{A time point on which the calculation is filtered}

\item{\code{df}}{The dataset the parameter or field is from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_blocks-remove_block"></a>}}
\if{latex}{\out{\hypertarget{method-table_blocks-remove_block}{}}}
\subsection{Method \code{remove_block()}}{
Remove block from the block data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_blocks$remove_block(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector specifying elements to remove from block data object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_blocks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-table_blocks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{table_blocks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
