% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{temp_col_rename}
\alias{temp_col_rename}
\title{Create new Generic Names for Columns with numeric table data}
\usage{
temp_col_rename(dat)
}
\arguments{
\item{dat}{a data frame}
}
\description{
Create new Generic Names for Columns with numeric table data
}
