% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tg_data.R
\docType{data}
\name{tg_data}
\alias{tg_data}
\title{Table Generator testing dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 10 columns.
}
\usage{
tg_data
}
\description{
Used to generate output for unit tests and compared to sas outputs
}
\keyword{datasets}
