% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{tg_gt}
\alias{tg_gt}
\title{Prepare the table generator data for output}
\usage{
tg_gt(tg_datalist, blockData, total_df, group)
}
\arguments{
\item{tg_datalist}{A list containing the data frames used to create the table}

\item{blockData}{The data frame for the construction of the blocks in the table or the block data object}

\item{total_df}{A data frame containing the totals by grouping variable}

\item{group}{A character denoting the grouping variable}
}
\value{
a data.frame containing output polished for presentation in `gt`
}
\description{
Prepare the data.frame so that it's ready for output via `gt` or other
}
\keyword{tabGen_repro}
