% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{tg_guide}
\alias{tg_guide}
\title{Table Generator Cicerone R6 Object}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 15.
}
\usage{
tg_guide
}
\description{
This object is used within the table generator module
to add help text to the various fields using 
a help buttom
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()}
}
\concept{tableGen Functions}
\keyword{datasets}
