% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{transpose_df}
\alias{transpose_df}
\title{transpose dataframes so they can all 
be used with rbind to generate
the gt tables}
\usage{
transpose_df(df, num)
}
\arguments{
\item{df}{the dataframe to transpose}

\item{num}{the number of rows to return}
}
\description{
transpose dataframes so they can all 
be used with rbind to generate
the gt tables
}
