% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{urin}
\alias{urin}
\title{Urinalysis PARAMCDs used to build STAN Table 41}
\format{
An object of class \code{character} of length 12.
}
\usage{
urin
}
\description{
Urinalysis PARAMCDs used to build STAN Table 41
}
\keyword{datasets}
