% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{writeJSON}
\alias{writeJSON}
\title{Write block data object to a JSON file}
\usage{
writeJSON(bd, file)
}
\arguments{
\item{bd}{A block data object}

\item{file}{File or connection to write to}
}
\description{
Write block data object to a JSON file
}
\keyword{table_blocks}
