% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-phuse_TDF_ADaM.R
\docType{data}
\name{adae}
\alias{adae}
\title{ADAE}
\format{
Data frame with 32,139 features and 34 fields
}
\source{
<https://github.com/phuse-org/TestDataFactory/blob/master/Updated/TDF_ADaM/adae.xpt>, downloaded 2020-06-17
}
\usage{
adae
}
\value{
a data.frame
}
\description{
Adverse Events Analysis Data from PHUSE Test Data Factory Project's GitHub.
}
\keyword{datasets}
