% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-phuse_TDF_ADaM.R
\docType{data}
\name{adlbc}
\alias{adlbc}
\title{ADLBC}
\format{
Data frame with 32,740 features and 58 fields
}
\source{
<https://github.com/phuse-org/TestDataFactory/blob/master/Updated/TDF_ADaM/adlbc.xpt>, downloaded 2020-06-17
}
\usage{
adlbc
}
\value{
a data.frame
}
\description{
Laboratory Results Chemistry Analysis Data from PHUSE Test Data Factory Project's GitHub.
}
\keyword{datasets}
