% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-phuse_TDF_ADaM.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL}
\format{
Data frame with 254 features and 51 fields
}
\source{
<https://github.com/phuse-org/TestDataFactory/blob/master/Updated/TDF_ADaM/adsl.xpt>, downloaded 2020-06-17
}
\usage{
adsl
}
\description{
Subject Level Analysis Dataset from Phuse Test Data Fractory Project's github.
}
\keyword{datasets}
