% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-phuse_TDF_ADaM.R
\docType{data}
\name{adtte}
\alias{adtte}
\title{ADTTE}
\format{
Data frame with 32,740 features and 58 fields
}
\source{
<https://github.com/phuse-org/TestDataFactory/blob/master/Updated/TDF_ADaM/adtte.xpt>, downloaded 2021-01-26
}
\usage{
adtte
}
\description{
Time to Event Analysis Dataset from Phuse Test Data Fractory Project's github.
}
\keyword{datasets}
