% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_heatCorr.R
\name{app_heatmap}
\alias{app_heatmap}
\title{IDEA heatmap plot of endpoint correlations}
\usage{
app_heatmap(
  data,
  yvar_x,
  yvar_y,
  time,
  value = "AVAL",
  cor_mthd = "pearson",
  show_sig = F,
  sig_level = 0.05
)
}
\description{
Create a line plot with a time variable as the x-axis
and using either the selected response variable
or if a PARAMCD is selected, then plot the corresponding value
to calculate the means. Lines are plotted by patient
}
\seealso{
Other popExp Functions: 
\code{\link{app_boxplot}()},
\code{\link{app_lineplot}()},
\code{\link{app_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{heatmap_srv}()},
\code{\link{heatmap_ui}()},
\code{\link{km_srv}()},
\code{\link{km_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
