% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_fct_heatCorr.R
\name{app_heatmap}
\alias{app_heatmap}
\title{tidyCDISC heatmap plot of endpoint correlations}
\usage{
app_heatmap(
  data,
  yvar_x,
  yvar_y,
  time,
  value = "AVAL",
  cor_mthd = "pearson",
  show_sig = F,
  sig_level = 0.05
)
}
\arguments{
\item{data}{Merged data to be used in plot}

\item{yvar_x}{character string of variable or paramcd names for x-axis}

\item{yvar_y}{character string of variable or paramcd names for y-axis}

\item{time}{character string of time variable}

\item{value}{character value: permitted values include "AVAL", "CHG", or
"BASE"}

\item{cor_mthd}{character string. Defaults to 'pearson' for calculating
correlation co-efficients. See `?cor` for more supported test stats.}

\item{show_sig}{logical, whether or not to display p-values for significant
tests statistics}

\item{sig_level}{dbl, defaulting to .05. Used to determine significance level
for correlation tests performed}
}
\description{
Create a heatmap comparing param/endpoints variable on both axis
}
\seealso{
Other popExp Functions: 
\code{\link{app_boxplot}()},
\code{\link{app_lineplot}()},
\code{\link{app_spaghettiplot}()}
}
\concept{popExp Functions}
\keyword{popEx}
