% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{cleanADAE}
\alias{cleanADAE}
\title{Function to clean and combine ADAE dataset with ADSL}
\usage{
cleanADAE(datafile, ADSL)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}
}
\description{
Function to clean and combine ADAE dataset with ADSL
}
