% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{col_for_list_expr}
\alias{col_for_list_expr}
\title{GT Column Names}
\usage{
col_for_list_expr(nm)
}
\arguments{
\item{nm}{A vector of column names}
}
\value{
A character object of class \code{from_markdown}.
}
\description{
The function creates the labels for each column using the total function so the columns are now NAME N= X
}
\keyword{tabGen_repro}
