% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{combineBDS}
\alias{combineBDS}
\title{Function to bind data rows from the list of user supplied data frames}
\usage{
combineBDS(datafile, ADSL)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}
}
\description{
Function to bind data rows from the list of user supplied data frames
}
