% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{common_rownames}
\alias{common_rownames}
\title{Identify Names of Columns}
\usage{
common_rownames(data, group)
}
\arguments{
\item{data}{the data to create columns with}

\item{group}{whether to group the data to calculate Ns}
}
\value{
A character vector
}
\description{
A function to transform the \code{gt} row names from generics to the column name and the total N of
each column
}
\keyword{tabGen_repro}
