% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{common_rownames}
\alias{common_rownames}
\title{transform the gt rownames
from generics to the column name
and the total N of each column}
\usage{
common_rownames(data, group)
}
\arguments{
\item{data}{the data to create columns with}

\item{group}{whether to group the data to calculate Ns}
}
\description{
transform the gt rownames
from generics to the column name
and the total N of each column
}
