% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{convertTGOutput}
\alias{convertTGOutput}
\title{Using the drag and drop blocks
and the shiny inputs,
create a dataframe of the columns 
and statistics to perform and add custom classes
to each of the column names based on their 
datafile of origin}
\usage{
convertTGOutput(agg, blocks)
}
\arguments{
\item{agg}{the aggregate statistic block 
to apply to the column}

\item{blocks}{the block corresponding 
to the column name to apply statistic on}
}
\description{
Using the drag and drop blocks
and the shiny inputs,
create a dataframe of the columns 
and statistics to perform and add custom classes
to each of the column names based on their 
datafile of origin
}
