% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{convertTGOutput}
\alias{convertTGOutput}
\title{Using the drag and drop blocks
and the shiny inputs,
create a dataframe of the columns 
and statistics to perform and add custom classes
to each of the column names based on their 
datafile of origin}
\usage{
convertTGOutput(agg, blocks)
}
\arguments{
\item{agg}{the aggregate statistic block 
to apply to the column}

\item{blocks}{the block corresponding 
to the column name to apply statistic on}
}
\description{
Using the drag and drop blocks
and the shiny inputs,
create a dataframe of the columns 
and statistics to perform and add custom classes
to each of the column names based on their 
datafile of origin
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_max_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_nested_freq}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
