% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{custom_class}
\alias{custom_class}
\title{Create custom S3 classes for each column block
based on the dataframe they came from}
\usage{
custom_class(x, df)
}
\arguments{
\item{x}{the name of each block}

\item{df}{the name of the data file
which will determine the class}
}
\description{
Create custom S3 classes for each column block
based on the dataframe they came from
}
