% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_methods.R
\name{custom_class}
\alias{custom_class}
\title{Create custom S3 classes for each column block
based on the dataframe they came from}
\usage{
custom_class(x, df)
}
\arguments{
\item{x}{the name of each block}

\item{df}{the name of the data file
which will determine the class}
}
\description{
Create custom S3 classes for each column block
based on the dataframe they came from
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{mean_summary}()}
}
\concept{tableGen Functions}
