% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{data_to_use_str}
\alias{data_to_use_str}
\title{Function to clean and combine ADAE dataset with ADSL}
\usage{
data_to_use_str(x, ae_data, bds_data)
}
\arguments{
\item{x}{string, naming a data.frame.}

\item{ae_data}{data.frame, of the AE variety}

\item{bds_data}{data.frame, of the BDS variety}
}
\value{
A `data.frame` object containing data of the AE variety if `x == "ADAE"` or one of the BDS variety if not.
}
\description{
Function to clean and combine ADAE dataset with ADSL
}
\keyword{tabGen_repro}
