% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{dropArea}
\alias{dropArea}
\title{The entire drop area of the blocks,
both statistical and columns}
\usage{
dropArea(name, id, ulid, class, styles, col)
}
\arguments{
\item{name}{the name of the drop area}

\item{id}{the id of the drop area}

\item{ulid}{the id of the list item}

\item{class}{the class of the drop area 
so that blocks can be dropped here but not others}

\item{styles}{the aesthetics to apply}

\item{col}{the width of the drop area}
}
\description{
The entire drop area of the blocks,
both statistical and columns
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
