% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{dropArea}
\alias{dropArea}
\title{The entire drop area of the blocks,
both statistical and columns}
\usage{
dropArea(name, id, ulid, class, styles, col)
}
\arguments{
\item{name}{the name of the drop area}

\item{id}{the id of the drop area}

\item{ulid}{the id of the list item}

\item{class}{the class of the drop area 
so that blocks can be dropped here but not others}

\item{styles}{the aesthetics to apply}

\item{col}{the width of the drop area}
}
\description{
The entire drop area of the blocks,
both statistical and columns
}
