% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_levels}
\alias{get_levels}
\title{Get Factor Levels}
\usage{
get_levels(x)
}
\arguments{
\item{x}{a vector}
}
\value{
x vector
}
\description{
Extracts the factor levels of a vector or returns the unique values if the vector is not a factor.
}
\references{
A character vector containing the levels of the factor/vector
}
\keyword{helpers}
