% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_mean.R
\name{IDEA_mean.ADSL}
\alias{IDEA_mean.ADSL}
\title{if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by}
\usage{
\method{IDEA_mean}{ADSL}(column, week, group = NULL, data)
}
\arguments{
\item{column}{the variable to perform summary statistics on,
this also contains the class of the column
based on the data file the column came from}

\item{week}{filter the variable by certain week}

\item{group}{the groups to compare for the summary statistics}

\item{data}{the data to use}
}
\value{
an summary statistic table of grouped variables
}
\description{
if ADSL supplied look for the column to take mean of
and look for a grouping variable to group_by
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
