% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popExp_km.R
\name{km_ui}
\alias{km_ui}
\title{Kaplan-Meier UI}
\usage{
km_ui(id, label = "km")
}
\arguments{
\item{id}{module ID}

\item{label}{module label}
}
\description{
This module contains the widgets needed to create
a KM Curve
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{km_srv}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{refact}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
