% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_meanSummary.R
\name{mean_summary}
\alias{mean_summary}
\title{the \code{mean_summary} function is used by
both the basic summary statistics block
and the change from baseline block}
\usage{
mean_summary(.data, to_count)
}
\arguments{
\item{.data}{is the data to pass to the summary statics}

\item{to_count}{is either AVAL or CHG}
}
\description{
the \code{mean_summary} function is used by
both the basic summary statistics block
and the change from baseline block
}
