% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataComply.R
\name{mod_dataComply_ui}
\alias{mod_dataComply_ui}
\title{dataComply UI Function}
\usage{
mod_dataComply_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
An empty shiny tagList
}
\description{
A shiny Module of a blank UI
}
\seealso{
Other dataComply Functions: 
\code{\link{gather_reqs}()},
\code{\link{gather_rules}()},
\code{\link{mod_dataComplyRules_server}()},
\code{\link{mod_dataComplyRules_ui}()},
\code{\link{mod_dataComply_server}()}
}
\concept{dataComply Functions}
