% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen.R
\name{mod_tableGen_server}
\alias{mod_tableGen_server}
\title{tableGen Server Function}
\usage{
mod_tableGen_server(input, output, session, datafile = reactive(NULL))
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{datafile}{all uploaded data files 
from the dataImport module}
}
\description{
tableGen Server Function
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe_column}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
