% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen.R
\name{mod_tableGen_server}
\alias{mod_tableGen_server}
\title{tableGen Server Function}
\usage{
mod_tableGen_server(
  input,
  output,
  session,
  datafile = reactive(NULL),
  filePaths = reactive(NULL)
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{datafile}{all uploaded data files 
from the dataImport module}
}
\description{
tableGen Server Function
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
