% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen.R
\name{mod_tableGen_server}
\alias{mod_tableGen_server}
\title{tableGen Server Function}
\usage{
mod_tableGen_server(input, output, session, datafile = reactive(NULL))
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{datafile}{all uploaded datafiles from the dataImport module}
}
\value{
the grouped blocks for columns
to be used in the tableGen UI
}
\description{
This module combines the BDS and ADSL files within the list
of dataframes to create a single data frame.
This data is what's ussed to create the
draggable blocks which appear in a drag zone,
and a custom shiny input binding returns the blocks in the drop zone 
as a list. This list is used to create a gt table output
which can be output as html oor csv.
In order to include OCCDS or custom data the structure of the import data
will need to be changed.
}
