% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen.R
\name{mod_tableGen_ui}
\alias{mod_tableGen_ui}
\title{tableGen UI Function}
\usage{
mod_tableGen_ui(id)
}
\arguments{
\item{id}{internal parameters for {shiny}.}
}
\description{
drag and drop table generator module
}
