% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{prep_adae}
\alias{prep_adae}
\title{Function to pre-filter the ADAE depending on the stan table selected}
\usage{
prep_adae(datafile, ADSL, input_recipe)
}
\arguments{
\item{datafile}{list of ADaM-ish dataframes}

\item{ADSL}{an ADSL data.frame}

\item{input_recipe}{The shiny input that keeps track of the recipe selected}
}
\description{
Function to pre-filter the ADAE depending on the stan table selected
}
\keyword{tabGen_repro}
