% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{prep_adsl}
\alias{prep_adsl}
\title{Function to pre-filter the ADSL depending on the stan table selected}
\usage{
prep_adsl(ADSL, input_recipe, stan_table_num)
}
\arguments{
\item{input_recipe}{The shiny input that keeps track of the recipe selected}

\item{data}{an ADSL}
}
\description{
Function to pre-filter the ADSL depending on the stan table selected
}
