% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{prep_adsl}
\alias{prep_adsl}
\title{Function to pre-filter the ADSL depending on the stan table selected}
\usage{
prep_adsl(ADSL, input_recipe)
}
\arguments{
\item{ADSL}{an ADSL data.frame}

\item{input_recipe}{The shiny input that keeps track of the recipe selected}
}
\description{
Function to pre-filter the ADSL depending on the stan table selected
}
\keyword{tabGen_repro}
