% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{prep_sas_table}
\alias{prep_sas_table}
\title{Organize SAS table into a format for comparing}
\usage{
prep_sas_table(data, as_is = FALSE, rm_desc_col = TRUE)
}
\arguments{
\item{data}{the sas_table dataframe, output from IDEA}

\item{as_is}{a logical which will determine if the table should be
minimally prepared for comparison (leaving collapsed columns together) or
prepared for optimal machine readability before comparison.}
}
\description{
Organize SAS table into a format for comparing
}
