% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SAScompare.R
\name{prep_tg_table}
\alias{prep_tg_table}
\title{Organize Table Generator table into a format for comparing}
\usage{
prep_tg_table(data, as_is = FALSE, num_dec = 1)
}
\arguments{
\item{data}{the tg_table dataframe, output from IDEA}

\item{as_is}{a logical which will determine if the table should be minimally
prepared for comparison (leaving collapsed columns together) or prepared
for optimal machine readability before comparison.}

\item{num_dec}{the number of desired decimal places on all numeric columns;
value will be passed to the 'digits' argument of the round() function}
}
\description{
Organize Table Generator table into a format for comparing
}
