% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{pretty_blocks}
\alias{pretty_blocks}
\title{Table Generator Pretty Block lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
pretty_blocks
}
\description{
This object is used within the table generator module
to add pretty names for each stat block when displayed in the table
}
\keyword{datasets}
