% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{readData}
\alias{readData}
\title{Read SAS Data Files}
\usage{
readData(study_directory, file_names)
}
\arguments{
\item{study_directory}{character, containing file path to CDISC data.frames
of interest}

\item{file_names}{list of CDISC data.frames}
}
\value{
A list of data frames containing the data from the SAS data files
}
\description{
Function to read the SAS list of user supplied data frames
}
\keyword{tabGen_repro}
