% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{readData}
\alias{readData}
\title{Function to read the SAS list of user supplied data frames}
\usage{
readData(study_directory, file_names)
}
\arguments{
\item{datalist}{list of CDISC dataframes}
}
\description{
Function to read the SAS list of user supplied data frames
}
