% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{recipe_column}
\alias{recipe_column}
\title{function to change the selected input of input$COLUMN based on the recipe block}
\usage{
recipe_column(input, dataset, column)
}
\arguments{
\item{input}{the shiny input id to change based on the recipe dropdown}

\item{dataset}{the data to use for all choices - ADSL reactive}

\item{column}{the specific column to select within the shiny input}
}
\description{
function to change the selected input of input$COLUMN based on the recipe block
}
