% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{recipe_column}
\alias{recipe_column}
\title{function to change the selected input of input$COLUMN based on the recipe block}
\usage{
recipe_column(input, dataset, column)
}
\arguments{
\item{input}{the shiny input id to change based on the recipe dropdown}

\item{dataset}{the data to use for all choices - ADSL reactive}

\item{column}{the specific column to select within the shiny input}
}
\description{
function to change the selected input of input$COLUMN based on the recipe block
}
\seealso{
Other tableGen Functions: 
\code{\link{CDISC_anova}()},
\code{\link{CDISC_chg}()},
\code{\link{CDISC_freq}()},
\code{\link{CDISC_mean.ADSL}()},
\code{\link{CDISC_mean.BDS}()},
\code{\link{CDISC_mean}()},
\code{\link{CDISC_methods}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{recipe}},
\code{\link{rowArea}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
