% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{refact}
\alias{refact}
\title{Refactor variables}
\usage{
refact(data, varc, varn)
}
\arguments{
\item{data}{A data frames}

\item{varc}{A character vector of variables names with character values}

\item{varn}{A character vector of variables names with numeric values}
}
\description{
Refactor variables by their common "N" counterparts
}
\seealso{
Other popExp Functions: 
\code{\link{IDEA_boxplot}()},
\code{\link{IDEA_heatmap}()},
\code{\link{IDEA_lineplot}()},
\code{\link{IDEA_spaghettiplot}()},
\code{\link{boxPlot_srv}()},
\code{\link{boxPlot_ui}()},
\code{\link{heatmap_srv}()},
\code{\link{heatmap_ui}()},
\code{\link{km_srv}()},
\code{\link{km_ui}()},
\code{\link{linePlot_srv}()},
\code{\link{linePlot_ui}()},
\code{\link{mod_popExp_server}()},
\code{\link{mod_popExp_ui}()},
\code{\link{scatterPlot_srv}()},
\code{\link{scatterPlot_ui}()},
\code{\link{spaghettiPlot_srv}()},
\code{\link{spaghettiPlot_ui}()}
}
\concept{popExp Functions}
