% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{rowArea}
\alias{rowArea}
\title{The entire row area of accordion blocks
generated within the table generator server}
\usage{
rowArea(bins, col)
}
\arguments{
\item{bins}{is the number of accordions to create}

\item{col}{is the column with to apply}
}
\description{
The entire row area of accordion blocks
generated within the table generator server
}
\seealso{
Other tableGen Functions: 
\code{\link{IDEA_anova}()},
\code{\link{IDEA_chg}()},
\code{\link{IDEA_freq}()},
\code{\link{IDEA_max_freq}()},
\code{\link{IDEA_mean.ADAE}()},
\code{\link{IDEA_mean.BDS}()},
\code{\link{IDEA_mean}()},
\code{\link{IDEA_methods}()},
\code{\link{IDEA_nested_freq}()},
\code{\link{IDEA_non_missing}()},
\code{\link{IDEA_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowBlock}()},
\code{\link{rowPallete}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
