% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_fct_blocks.R
\name{rowBlock}
\alias{rowBlock}
\title{Create a single draggable block for each column}
\usage{
rowBlock(name, classname)
}
\arguments{
\item{name}{the column name}

\item{classname}{the dataset of the column}
}
\description{
Create a single draggable block for each column
}
\seealso{
Other tableGen Functions: 
\code{\link{app_anova}()},
\code{\link{app_chg}()},
\code{\link{app_freq}()},
\code{\link{app_max_freq}()},
\code{\link{app_mean.ADAE}()},
\code{\link{app_mean.BDS}()},
\code{\link{app_mean}()},
\code{\link{app_methods}()},
\code{\link{app_nested_freq}()},
\code{\link{app_non_missing}()},
\code{\link{app_y}()},
\code{\link{convertTGOutput}()},
\code{\link{custom_class}()},
\code{\link{dropArea}()},
\code{\link{mean_summary}()},
\code{\link{mod_tableGen_server}()},
\code{\link{mod_tableGen_ui}()},
\code{\link{rowArea}()},
\code{\link{rowPallete}()},
\code{\link{std_err}()},
\code{\link{tg_guide}}
}
\concept{tableGen Functions}
