% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{setGroup}
\alias{setGroup}
\title{Set the title for the table object}
\usage{
setGroup(bd, group_by)
}
\arguments{
\item{bd}{A block data object}

\item{group_by}{A field to group the table by}
}
\value{
The \code{bd} block data object with updated group by field
}
\description{
Set the title for the table object
}
\examples{

datalist <- list(ADSL = tidyCDISC::adsl, ADVS = tidyCDISC::advs, 
                 ADAE = tidyCDISC::adae, ADLBC = tidyCDISC::adlbc)
bd <- createBlockdata(datalist)
setGroup("TRT01P")
bd$group_by
}
\keyword{table_blocks}
