% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{setTitle}
\alias{setTitle}
\title{Set the title for the table object}
\usage{
setTitle(bd, title)
}
\arguments{
\item{bd}{A block data object}

\item{title}{The title for the table}
}
\value{
The \code{bd} block data object with supplied title
}
\description{
Set the title for the table object
}
\examples{

datalist <- list(ADSL = tidyCDISC::adsl, ADVS = tidyCDISC::advs, 
                 ADAE = tidyCDISC::adae, ADLBC = tidyCDISC::adlbc)
bd <- createBlockdata(datalist)
setTitle("Table 1")
bd$title
}
\keyword{table_blocks}
