% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\name{tg_gt}
\alias{tg_gt}
\title{Create the gt table object for TG}
\usage{
tg_gt(tg_datalist, blockData, total_df, group)
}
\arguments{
\item{tg_datalist}{A list containing the data frames used to create the table}

\item{blockData}{The data for the construction of the blocks in the table}

\item{total_df}{A data frame containing the totals by grouping variable}

\item{group}{A character denoting the grouping variable}
}
\description{
A wrapper for other functions to create the `gt` object from the data
}
