% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableGen_utils.R
\docType{data}
\name{tg_guide}
\alias{tg_guide}
\title{Table Generator Cicerone R6 Object}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 13.
}
\usage{
tg_guide
}
\description{
This object is used within the table generator module
to add help text to the various fields using 
a help buttom
}
\keyword{datasets}
